create useLinkValidation.ts custom hook to abstract some of the code for readability
